requirements.txt
altgraph==0.17.2
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect==1.1.0
future==0.18.2
idna==3.3
lxml==4.8.0
pefile==2021.9.3
Pillow==9.1.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pywin32-ctypes==0.2.0
requests==2.27.1
urllib3==1.26.9
